--- helpers/Photos.ts
@@ -1,62 +1,0 @@
-const breakpoints = [1080, 900, 384, 256, 128, 96, 64, 48];
-
-const unsplashLink = (id: string, width: number, height: number) =>
-    `https://source.unsplash.com/${id}/${width}x${height}`;
-
-// const unsplashPhotos = [
-//     { id: "Osq7UAVxIOI", width: 1080, height: 780 },
-//     { id: "Dhmn6ete6g8", width: 1080, height: 1620 },
-//     { id: "RkBTPqPEGDo", width: 1080, height: 720 },
-//     { id: "Yizrl9N_eDA", width: 1080, height: 721 },
-//     { id: "KG3TyFi0iTU", width: 1080, height: 1620 },
-//     { id: "Jztmx9yqjBw", width: 1080, height: 607 },
-//     { id: "-heLWtuAN3c", width: 1080, height: 608 },
-//     { id: "xOigCUcFdA8", width: 1080, height: 720 },
-//     { id: "1azAjl8FTnU", width: 1080, height: 1549 },
-//     { id: "ALrCdq-ui_Q", width: 1080, height: 720 },
-//     { id: "twukN12EN7c", width: 1080, height: 694 },
-//     { id: "9UjEyzA6pP4", width: 1080, height: 1620 },
-//     { id: "sEXGgun3ZiE", width: 1080, height: 720 },
-//     { id: "S-cdwrx-YuQ", width: 1080, height: 1440 },
-//     { id: "q-motCAvPBM", width: 1080, height: 1620 },
-//     { id: "Xn4L310ztMU", width: 1080, height: 810 },
-//     { id: "ls94iFAQerE", width: 1080, height: 1620 },
-//     { id: "X48pUOPKf7A", width: 1080, height: 160 },
-//     { id: "GbLS6YVXj0U", width: 1080, height: 810 },
-//     { id: "9CRd1J1rEOM", width: 1080, height: 720 },
-//     { id: "xKhtkhc9HbQ", width: 1080, height: 1440 },
-// ];
-
-const unsplashPhotos = [
-    { key:'1', src: "./images/album/wd1.jpg", width: 900, height: 1400 },
-    { key:'2', src: "./images/album/wd2.jpg", width: 900, height: 1200 },
-    { key:'3', src: "./images/album/wd3.jpg", width: 900, height: 1200 },
-    { key:'4', src: "./images/album/wd4.jpg", width: 900, height: 1200 },
-    { key:'5', src: "./images/album/wd5.jpg", width: 900, height: 1200 },
-    { key:'6', src: "./images/album/wd6.jpg", width: 900, height: 1200 },
-    { key:'7', src: "./images/album/wd7.jpg", width: 900, height: 1200 },
-    { key:'8', src: "./images/album/wd8.jpg", width: 900, height: 1200 },
-    { key:'9', src: "./images/album/wd9.jpg", width: 900, height: 1200 },
-    { key:'10', src: "./images/album/wd10.jpg", width: 900, height: 1400 },
-    { key:'11', src: "./images/album/wd11.jpg", width: 900, height: 1200 },
-    { key:'12', src: "./images/album/wd12.jpg", width: 900, height: 1200 },
-    { key:'13', src: "./images/album/wd13.jpg", width: 900, height: 1200 },
-    { key:'14', src: "./images/album/wd14.jpg", width: 900, height: 1000 },
-];
-
-
-const photos = unsplashPhotos.map((photo) => ({
-    src: photo.src,
-    width: photo.width,
-    height: photo.height,
-    images: breakpoints.map((breakpoint) => {
-        const height = Math.round((photo.height / photo.width) * breakpoint);
-        return {
-            src: photo.src,
-            width: breakpoint,
-            height,
-        };
-    }),
-}));
-
-export default photos;
