+++ types/navigation-props.ts
@@ -1,0 +1,5 @@
+export type NavigationProps = {
+    open?: boolean
+    onOpen?: (val:boolean) => void,
+    onAudio?: (val:string) => void
+}
