+++ types/scroll-position-props.ts
@@ -1,0 +1,3 @@
+export type ScrollPositionProps = {
+    yPosition :number
+}
